More Refactored Files
