+++ data-query/spotbugs-excludes.xml
@@ -27,8 +27,8 @@
       <Class name="gov.va.api.health.dataquery.service.controller.allergyintolerance.DatamartAllergyIntolerance"/>
       <Class name="gov.va.api.health.dataquery.service.controller.allergyintolerance.DatamartAllergyIntolerance$Reaction"/>
       <Class name="gov.va.api.health.dataquery.service.controller.diagnosticreport.DatamartDiagnosticReport"/>
-      <Class name="gov.va.api.health.dataquery.service.controller.diagnosticreport.DatamartDiagnosticReports"/>
-      <Class name="gov.va.api.health.dataquery.service.controller.diagnosticreport.DatamartDiagnosticReports$DiagnosticReport"/>
+      <Class name="gov.va.api.health.dataquery.service.controller.diagnosticreport.v1.DatamartDiagnosticReports"/>
+      <Class name="gov.va.api.health.dataquery.service.controller.diagnosticreport.v1.DatamartDiagnosticReports$DiagnosticReport"/>
       <Class name="gov.va.api.health.dataquery.service.controller.medicationorder.DatamartMedicationOrder"/>
       <Class name="gov.va.api.health.dataquery.service.controller.observation.DatamartObservation"/>
       <Class name="gov.va.api.health.dataquery.service.controller.organization.DatamartOrganization"/>
